Add an optional github-token input
Doing it this way means we can default it to the action's token; then
one you almost always want. When we were accepting it in env, users
always had to specify it. This was implemented that way due to lack of
experience with actions only.

We keep a small conditional so that any users specifying it via env will
continue to function as before (and we can avoid a major bump).