Fixup some bugs not caught on CI
Our "check for existence" API endpoint was wrong, so we never saw
something as existing.

Our "create refs" body was wrong, so even if we fixed the
existence-check, it would still never find what we created.

Both are fixed and the test tag is bumped. Additional CI has been added.